fix(builder): remove duplicated babel-plugin-import
