+++ packages/cli/babel-preset-app/src/index.ts
@@ -24,6 +24,7 @@ const defaultOptions: Options = {
   useTsLoader: false,
   lodash: {},
   styledComponents: {},
+  importAntd: true,
 };
 
 export const getBabelConfig = (options?: Options): BabelOptions => {
