+++ packages/toolkit/utils/src/resolve.ts
@@ -34,3 +34,15 @@ export const isPackageInstalled = (
     return false;
   }
 };
+
+export const getAntdMajorVersion = (appDirectory: string) => {
+  try {
+    const pkgJsonPath = require.resolve('antd/package.json', {
+      paths: [appDirectory],
+    });
+    const { version } = require(pkgJsonPath);
+    return Number(version.split('.')[0]);
+  } catch (err) {
+    return null;
+  }
+};
